simplification of http calls, updates to readme, rename top level folder
