+++ LittleFs_RemoteDisk/src/HighLevelApp/curlFunctions.h
