--- Littlefs/src/HighLevelApp/httpGet.h
@@ -1,7 +1,0 @@
-/* Copyright (c) Microsoft Corporation. All rights reserved.
-   Licensed under the MIT License. */
-#include <stdint.h>
-
-#pragma once
-
-uint8_t* readBlockData(uint32_t offset, uint32_t size);
